feat: Set Turkey as default country on customer add page
This commit updates the "Add Customer" page to automatically select Turkey ("TR") as the default country when the page loads. If Turkey is not found in the country list, it will default to the first available country. This improves the user experience for users in Turkey by pre-filling the country field.